requirements.txt
django==3.0.7
django-admin-sortable2==0.7.8
djangorestframework==3.12.2
django-debug-toolbar==3.2
dj-database-url==0.5.0
geopy==2.1.0
markdown==3.3.3
Pillow==8.1.0
requests==2.22.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.18
environs==9.3.1
wemake-python-styleguide==0.14.1
